+++ MikroSqlDbYedek.Win/Theme/ModernToolStripRenderer.cs
@@ -39,8 +39,8 @@ namespace MikroSqlDbYedek.Win.Theme
                 ModernTheme.SetHighQuality(g);
 
                 var bgColor = e.Item.Pressed
-                    ? Color.FromArgb(30, 0, 0, 0)
-                    : Color.FromArgb(15, 0, 0, 0);
+                    ? Color.FromArgb(40, 255, 255, 255)
+                    : Color.FromArgb(20, 255, 255, 255);
 
                 using (var path = ModernTheme.CreateRoundedRectanglePath(rect, 4))
                 using (var brush = new SolidBrush(bgColor))
@@ -83,7 +83,7 @@ namespace MikroSqlDbYedek.Win.Theme
             if (e.Item.Selected && e.Item.Enabled)
             {
                 var rect = new Rectangle(2, 1, e.Item.Width - 4, e.Item.Height - 2);
-                using (var brush = new SolidBrush(Color.FromArgb(15, 0, 120, 212)))
+                using (var brush = new SolidBrush(Color.FromArgb(25, 0, 150, 80)))
                 {
                     e.Graphics.FillRectangle(brush, rect);
                 }
@@ -109,12 +109,12 @@ namespace MikroSqlDbYedek.Win.Theme
         public override Color ImageMarginGradientEnd => ModernTheme.SurfaceColor;
         public override Color SeparatorDark => ModernTheme.DividerColor;
         public override Color SeparatorLight => ModernTheme.SurfaceColor;
-        public override Color MenuItemSelected => Color.FromArgb(15, 0, 120, 212);
+        public override Color MenuItemSelected => Color.FromArgb(25, 0, 150, 80);
         public override Color MenuItemBorder => Color.Transparent;
-        public override Color ButtonSelectedHighlight => Color.FromArgb(15, 0, 120, 212);
-        public override Color ButtonSelectedGradientBegin => Color.FromArgb(15, 0, 120, 212);
-        public override Color ButtonSelectedGradientEnd => Color.FromArgb(15, 0, 120, 212);
-        public override Color ButtonPressedGradientBegin => Color.FromArgb(30, 0, 120, 212);
-        public override Color ButtonPressedGradientEnd => Color.FromArgb(30, 0, 120, 212);
+        public override Color ButtonSelectedHighlight => Color.FromArgb(20, 255, 255, 255);
+        public override Color ButtonSelectedGradientBegin => Color.FromArgb(20, 255, 255, 255);
+        public override Color ButtonSelectedGradientEnd => Color.FromArgb(20, 255, 255, 255);
+        public override Color ButtonPressedGradientBegin => Color.FromArgb(40, 255, 255, 255);
+        public override Color ButtonPressedGradientEnd => Color.FromArgb(40, 255, 255, 255);
     }
 }